eiffeltest timeboxing WIP: without timeout, it works at last.

+++ src/smarteiffel/eiffeltest/eiffeltest_server_process.e
@@ -1,32 +1,32 @@
 -- This file is part of Liberty Eiffel The GNU Eiffel Compiler Tools and Libraries.
 -- See the Copyright notice at the end of this file.
 --
-expanded class EIFFELTEST_SERVER_PROCESS
+class EIFFELTEST_SERVER_PROCESS
 
 insert
    LOGGING
       redefine
-         default_create, copy, is_equal
+         copy, is_equal
       end
    PROCESS_WAIT
       rename
          in as waitpid_in,
          job as waitpid_job
       redefine
-         default_create, copy, is_equal
+         copy, is_equal
+      end
+   RECYCLABLE
+      redefine
+         copy, is_equal
       end
 
-feature {}
-   default_create is
-      do
-         process := Void
-         cleanup := Void
-         cmd := Void
-      end
+create {EIFFELTEST_SERVER_RUN_TESTS}
+   default_create
 
 feature {ANY}
    copy (other: like Current) is
       do
+         port := other.port
          process := other.process
          cleanup := other.cleanup
          cmd := other.cmd
@@ -34,6 +34,9 @@ feature {ANY}
 
    is_equal (other: like Current): BOOLEAN is
       do
+         check
+            same_process: port = other.port
+         end
          Result := process = other.process
             and then cleanup = other.cleanup
             and then cmd = other.cmd
@@ -55,16 +58,9 @@ feature {EIFFELTEST_SERVER_RUN_TESTS}
       end
 
    on_done (a_status: INTEGER) is
-      local
-         status: INTEGER
       do
-         log.info.put_line(once "Server #(1): timeboxed command finished: #(2)" # port.out # cmd)
          process.wait
-         status := process.status
-         check
-            process.is_finished
-            a_status = status
-         end
+         log.info.put_line(once "Server #(1): timeboxed command finished (status #(3)): #(2)" # port.out # cmd # a_status.out)
          if cleanup /= Void then
             cleanup.call([a_status])
          end
@@ -72,22 +68,32 @@ feature {EIFFELTEST_SERVER_RUN_TESTS}
 
    on_timeout is
       local
-         pid, status: INTEGER
+         mypid, pid, status: INTEGER
       do
-         log.info.put_line(once "Server #(1): timeboxed command timed out: #(2)" # port.out # cmd)
-         --| **** TODO: cross-platform
-         pid := process.id
-         c_inline_h("#include <signal.h>%N")
-         c_inline_c("kill(_pid, 15);%N")
-         log.info.put_line(once "Server #(1): timeboxed command killed: #(2)" # port.out # pid.out)
-         process.wait
-         status := process.status
+         if process.is_finished then
+            status := process.status
+            log.info.put_line(once "Server #(1): timeboxed command timed out but finished (status #(3)): #(2)" # port.out # cmd # status.out)
+         else
+            log.info.put_line(once "Server #(1): timeboxed command timed out: #(2)" # port.out # cmd)
+            --| **** TODO: cross-platform
+            c_inline_c("_mypid = getpid();%N")
+            pid := process.id
+            log.trace.put_line(once "Server #(1): killing timeboxed command: #(2) (#(3))" # port.out # pid.out # mypid.out)
+            if pid /= mypid then
+               c_inline_h("#include <signal.h>%N")
+               c_inline_c("kill(_pid, 15);%N")
+               process.wait
+               status := process.status
+            else
+               log.warning.put_line(once "Server #(1): pid is 0!!" # port.out)
+            end
+         end
          if cleanup /= Void then
             cleanup.call([-1])
          end
       end
 
-   set (a_port: like port; a_cmd: like cmd; a_cleanup: like cleanup) is
+   set (a_port: like port; a_timeout: like timeout; a_cmd: like cmd; a_cleanup: like cleanup) is
       require
          a_port > 0
          not is_set
@@ -95,10 +101,12 @@ feature {EIFFELTEST_SERVER_RUN_TESTS}
       do
          log.info.put_line(once "Server #(1): preparing timeboxed command: #(2)" # a_port.out # a_cmd)
          port := a_port
+         timeout := a_timeout
          cmd := a_cmd
          cleanup := a_cleanup
       ensure
          port = a_port
+         timeout = a_timeout
          cmd = a_cmd
          cleanup = a_cleanup
          is_set
@@ -110,13 +118,19 @@ feature {EIFFELTEST_SERVER_RUN_TESTS}
          is_set
          not is_running
       local
-         split: FAST_ARRAY[STRING]
+         split: RING_ARRAY[STRING]; c: STRING
+         process_factory: PROCESS_FACTORY
       do
          log.info.put_line(once "Server #(1): now running timeboxed command: #(2)" # port.out # cmd)
-         create split.make(0)
+         create split.make(1, 0)
          cmd.split_in(split)
-         waitpid_job.trigger(1000) --| **** intentionally short for test, obviously needs to be changed
-         process := process_factory.execute(split.first, split)
+         waitpid_job.trigger(timeout)
+         c := split.first
+         split.remove_first
+         process_factory.set_direct_input(True)
+         process_factory.set_direct_output(True)
+         process_factory.set_direct_error(True)
+         process := process_factory.execute(c, split)
       ensure
          is_running
       end
@@ -138,12 +152,22 @@ feature {EIFFELTEST_SERVER_PROCESS}
    process: PROCESS
    cleanup: PROCEDURE[TUPLE[INTEGER]]
    port: INTEGER
+   timeout: INTEGER
 
-   process_factory: PROCESS_FACTORY is
+   process_factory_: PROCESS_FACTORY is
       once
          Result.set_direct_input(True)
          Result.set_direct_output(True)
          Result.set_direct_error(True)
+      end
+
+feature {RECYCLING_POOL}
+   recycle is
+      do
+         cmd := Void
+         process := Void
+         cleanup := Void
+         port := 0
       end
 
 end -- class EIFFELTEST_SERVER_PROCESS
